stage3: Add perl-threads RPM.
Because the stage3 chroot is built from the host, and the host has a
separate perl-threads RPM which installs in a different place
(/usr/lib64/perl5/vendor_perl/threads.pm vs
/usr/lib64/perl5/threads.pm), we get a versioning error when trying to
use the threads package.

threads object version 2.07 does not match bootstrap parameter 2.09 at /usr/lib64/perl5/DynaLoader.pm line 210.

This happens because it's using the host-copied
/usr/lib64/perl5/vendor_perl/threads.pm instead of
/usr/lib64/perl5/threads.pm.

Anyway the solution is to build perl-threads and install it.  Since it
has a higher version it will override the perl-built package of the
same name, and it will overwrite the host version.

All will be fixed properly in stage4 ...