stage4-build: Fix umount.
Because of the way we build the stage3, every file in the fiesystem is
owned by a non-root user (eg. for me it is 1000.1000).

Because umount is a setuid program, this means umount runs as UID 1000
(instead of root).

chown umount to root.root.  This probably loses the setuid flag in the
process, but that doesn't matter because we are only going to run it
as root.