Add systemtap-sdt-devel.
This is a cheat.  This package is normally archful, but it contains
only noarch files.  To generate this file I simply modified the
systemtap spec file to add BuildArch: noarch and remove an unnecessary
dependency (see patch below).  This allows us to build a
systemtap-sdt-devel.noarch package on x86_64.

I believed that this could be used to satisfy the perl-devel
dependency (and indirect dependency of a lot of Perl packages) in
stage4.  Unfortunately there's still an extra dependency needed:
perl-devel -> glibc-devel -> libgcc(riscv64), so we still need GCC for
this.

However once we have GCC it will allow perl-devel to be installed so
it's still an improvement.

diff --git a/systemtap.spec b/systemtap.spec
index 23ad6b8..13bf8da 100644
--- a/systemtap.spec
+++ b/systemtap.spec
@@ -274,13 +274,14 @@ Summary: Static probe support tools
 Group: Development/System
 License: GPLv2+ and Public Domain
 URL: http://sourceware.org/systemtap/
-%if %{with_pyparsing}
-%if %{with_python3}
-Requires: python3-pyparsing
-%else
-Requires: pyparsing
-%endif
-%endif
+BuildArch: noarch
+#%if %{with_pyparsing}
+#%if %{with_python3}
+#Requires: python3-pyparsing
+#%else
+#Requires: pyparsing
+#%endif
+#%endif

 %description sdt-devel
 This package includes the <sys/sdt.h> header file used for static